+++ SOPT-iOS/Projects/Features/HomeFeature/Sources/HomeScene/ViewModel/HomeForMemberViewModel.swift
@@ -46,6 +46,19 @@ struct GroupInfo {
     let imageURL: String
 }
 
+struct CoffeeChatHostInfo {
+    let memberId: Int
+    let bio: String
+    let topicTypeList: [String]
+    let profileImage: String?
+    let name: String
+    let career: String?
+    let organization: String
+    let companyJob: String?
+    let soptActivities: [String]
+    let nowActivity: String?
+}
+
 public class HomeForMemberViewModel: HomeForMemberViewModelType {
     
     // MARK: - Properties
@@ -76,6 +89,13 @@ public class HomeForMemberViewModel: HomeForMemberViewModelType {
         GroupInfo(title: "모임 타이틀이고 두 줄이 넘어가면 줄어들어야 합니다", category: .study, canJoinOnlyActiveGeneration: true, joinableParts: ["안드로이드"], canJoinAllParts: false, status: .recruitmentComplete, imageURL: "https://www.petmove.co.kr/content/images/size/w2400/2023/09/ying-zhu-4UZfmxvc5Qk-unsplash.jpg"),
         GroupInfo(title: "모임 타이틀이고 두 줄이 넘어가면 줄어들어야 합니다", category: .event, canJoinOnlyActiveGeneration: true, joinableParts: ["iOS"], canJoinAllParts: false, status: .applyAble, imageURL: "https://flexible.img.hani.co.kr/flexible/normal/960/960/imgdb/resize/2019/0121/00501111_20190121.JPG"),
         GroupInfo(title: "모임 타이틀이고 두 줄이 넘어가면 줄어들어야 합니다", category: .study, canJoinOnlyActiveGeneration: true, joinableParts: ["iOS"], canJoinAllParts: false, status: .applyAble, imageURL: "https://flexible.img.hani.co.kr/flexible/normal/960/960/imgdb/resize/2019/0121/00501111_20190121.JPG")
+    ]
+    
+    // TODO: 서버 연결 필요
+    let coffeeChatHostInfoList: [CoffeeChatHostInfo] = [
+        CoffeeChatHostInfo(memberId: 0, bio: "디자인 관련 고민이 있다면, 함께 나눠봐요!", topicTypeList: ["커리어", "면접", "포트폴리오"], profileImage: "https://i.pinimg.com/736x/d0/1e/78/d01e78f19a709a859f7c23d1cab11db3.jpg", name: "재영이", career: "주니어(0-3년차)", organization: "Google", companyJob: "Product Designer", soptActivities: ["29기 디자인"], nowActivity: "35기 웹"),
+        CoffeeChatHostInfo(memberId: 0, bio: "디자인 관련 고민이 있다면, 함께 나눠봐요!", topicTypeList: ["커리어", "면접", "포트폴리오"], profileImage: "https://i.pinimg.com/736x/97/08/4c/97084c4f037ac2db897535268ca475b3.jpg", name: "포차코", career: "주니어(0-3년차)", organization: "Google", companyJob: "Product Designer", soptActivities: ["29기 디자인"], nowActivity: "35기 웹"),
+        CoffeeChatHostInfo(memberId: 0, bio: "디자인 관련 고민이 있다면, 함께 나눠봐요!", topicTypeList: ["커리어", "면접", "포트폴리오"], profileImage: "https://i.pinimg.com/736x/d0/1e/78/d01e78f19a709a859f7c23d1cab11db3.jpg", name: "차은우", career: "주니어(0-3년차)", organization: "Google", companyJob: "Product Designer", soptActivities: ["29기 디자인"], nowActivity: "35기 웹")
     ]
     
     // MARK: - Inputs
